made refactoring changes in packaging naming conventions
