+++ src/main/java/com/mg105/use_cases/Inventory/InventoryInteractor.java
@@ -1,13 +1,11 @@
-package com.mg105.usecase.Inventory;
+package com.mg105.use_cases.Inventory;
 
 import com.mg105.entities.BattleCharacter;
 import com.mg105.entities.GameState;
 import com.mg105.outputds.ItemDetails;
-import com.mg105.presenterinterfaces.InventoryPresenterInterface;
+import com.mg105.presenter_interfaces.InventoryPresenterInterface;
 import com.mg105.utils.ItemConstants;
 import com.mg105.utils.PartyConstants;
-
-import java.util.NoSuchElementException;
 
 /**
  * This class is the only class that should directly interact with the inventory
