+++ src/test/java/com/mg105/usecase/Inventory/InventoryInteractorTest.java
@@ -7,7 +7,8 @@ import com.mg105.entities.Move;
 import com.mg105.entities.items.HealthPotion;
 import com.mg105.entities.items.UpgradeToken;
 import com.mg105.outputds.ItemDetails;
-import com.mg105.presenterinterfaces.InventoryPresenterInterface;
+import com.mg105.presenter_interfaces.InventoryPresenterInterface;
+import com.mg105.use_cases.Inventory.InventoryInteractor;
 import com.mg105.utils.ItemConstants;
 import org.junit.jupiter.api.Test;
